Support per-file sync for streams created using the FileFactory.
This is done by opening the file using the POSIX open() call and
associating it with a streambuf. Depending on which C++ library
implementation is used, either an internal open call is used or a
non-standard constructor call. While it is a pity to having to resort to
an internal call for this, knowing the file handle enables using the
fsync() call when calling sync() on streambuf.

Issue: SWP-168452
GIT_ORIGIN_SPP_REV_ID: b20247d13c5fee4b601edb8af1e5bc1658bfb9bf