Add ownership control option to JsonWriter
Issue: SWP-202773

When using JsonWriter on the kSynced mode, a temporary file is created
and then moved to be the final file if the operation succeeds.
It might be needed to set the ownership of the temporary file to reflect
that of the original file (thus having a transparent update process).
So far the options for controlling this behavior were not exposed to
JsonWriter. This PR exposes the options as well as explaining when and
how to use them.

GIT_ORIGIN_SPP_REV_ID: a9d33e5c2a0e239046b5092c48a9f5d21fd0599d