Work towards compliance to clang warning 'Wunused-template'
In ipnext scope we want to establish using compiler warning `Wunused-template`
to enforce MISRA C++:2023 rules 0.2.1 and 0.2.4 (variables/functions with
limited visibility should be used at least once).
While testing the rollout of this warning we discovered some issue in this
project we attempt to fix in this change.
Please note, it seems this compiler warning is confused by static template
functions in header files and reports them falsely as unused.
Although, this seems to be an error the the compiler warning, we still
change the code. The static keyword makes no sense for a template function in a
header file and thus this is a trivial non functional fix.

Issue: DCS-12506
GIT_ORIGIN_SPP_REV_ID: 7f69024d582dede1b0fe32713133c7ec34a1bc39