Change default to kUseTargetFileUID
Issue: SWP-202773

Based on team feedback, the default behavior will be setting the UID and
GID to be the ones from the target file.
If that is not appropriate for an use-case, the developer will have to
be aware of the possible permission issue, and use kUseCurrentProcessUID
instead on his/her code.
That use-case is documented on the API's decription, so it should be
easy for him/her to figure out how to accomplish the desired behavior.

GIT_ORIGIN_SPP_REV_ID: f94f3568c1c0632886fbbf193260eadf840a4e65