Merge pull request #1608 from swh/account_for_misra23_0_1_2
Explicitly discard return value (account for misra23 0.1.2

After an analysis of the return values I've done the following:

If the header was an impl-detail header: add include-dependency to amp::ignore to explicitly discard the return value. Only if the return-type is different between C++14 and C++17 (std::vector::emplace_back I've explicitly cast it to void because amp::ignore cannot swallow void return values (see C++14 amp-CI-configuration)
if the header was part of the public api and it already had an include-dependency to amp::ignore reuse it
if the header was part of the public api and it had not an include-dependency to amp::ignore use a static-cast to void to explicity discard the return value while keeping the "required symbols" as narrow as possible.

GIT_ORIGIN_FUTURE_CPP_REV_ID: 6a4cb9bd4e6cc5d247a032b17ea2a952a88df7b6