safecpp: Fix compiler warning in QNX8
The warning appears because in QNX8 the type changed and there is a
type mismatch. With this change we use our comparison functions that
behave similiar like the standard cmp_* and in this way we do not
have to deal with all type combinations. Complexity is handled by the
CmpLessEqual function.

Issue: SWP-203327
GIT_ORIGIN_SPP_REV_ID: 94bcae8d3852bcb4ec0e2122ffbbd8a760199c21