Merge pull request #1625 from swh/tr_export_static_const_from_utility
Export static_const.hpp directly from amp_utility.hpp

I was wondering if static_const is public. I found that it is not in a detail namespace, but in a private/utility header. Various headers include it and thereby provide it transitively, which makes it available from amp_utility.hpp as well. But it is not explicitly included and exported there, which is what this PR intends to address.

GIT_ORIGIN_FUTURE_CPP_REV_ID: ac3b72e7a8e328e44abe450e474d46564e66f001