bmw::Result bindings for Rust
Introduce a binary compatible version for bmw::Result and associated
types. These include

* bmw::details::expected
* std::string_view
* bmw::Error
* bmw::ErrorDomain

The result can be transferred across FFI boundaries and converted into a
Rust std::result::Result so that it can be treated as usual in Rust and
to leverage the ecosystem around this type.

If one wants to use the Result FFI bindings with CXX, there is a macro
that can be used to import typedefs on C++ side.

GIT_ORIGIN_SPP_REV_ID: d1171eedfadac351f1cf6cf86d2d23e20bc14d77