filesystem: Remove superflous call to remove
remove calls unlink for files and rmdir for directories. Here we are
dealing with a file and the call to unlink is already done.
This call does not bring any value and it is just confusing.

Issue: SWP-218363
GIT_ORIGIN_SPP_REV_ID: 4147c0a940a29d752c1f517592346561a62a34b3