filesystem: Fix API internal method
It is better to just pass a string_view because we do not need to copy
and also not story/consume it. We just need to read the value.
This also removes a coverity issue that complains about additional
copy.

Issue: SWP-218363
GIT_ORIGIN_SPP_REV_ID: 2c747d274d7488d4747a54cbe82b316992b8143d