aas/analysis: auto-fix non-compliant header guards
With the migration to the safe-posix-platform repository, the clang-tidy
header guard checks turned out to become dysfunctional. Reason for that
is a restriction in the checker's verification logic for the `DDAD` guard
style. This got fixed with the 19.1.7-6 release of the llvm toolchain.

The fixes of this commit got auto-generated via the following command:
```
bazel run //platform/aas/quality/clang_tidy:fix_extra_checks
```

GIT_ORIGIN_SPP_REV_ID: 51132f253df84984d3194a558bd933611ddb4d30